Use correct method names for memcached module
The memcached module provides aliases increment and decrement for incr
and decr but they are not documented and not part of the TypeScript type
definitions.

This updates the rate limit store to use the documented method names.
This means the rate limiter code can be more easily tested using a mock
memcached client.